Adds ETL process for camps and events.
An ETL process is added in class EtlService. It queries the Playa
Events API for JSON about camps and events, then parses that JSON into
documents which are stored in Elasticsearch. The service executes once
a day at 1 am, but an admin endpoint for manually triggering the
operation is also provided.

Side Effects:
+ New service class EtlService carrying most of the logic.
+ New DTO classes for information retrieved from playa events.
+ New model classes (documents) for storage in Elasticsearch.
+ New resource exposing admin endpoint.
+ Build now separates slow integration tests from fast ones. Slower
tests run in integration-test phase of maven build.

Testing:
+ New unit test for ETL process.
+ New integration test that hits the Playa events API.