Adds random full-screen background image to site.
Modifies index.html for the app so that a full-screen background image
(minus header and footer) is displayed. This was done as part of a
series of changes to establish a base for the UI. The background image
is chosen at random from a small set of images added with this commit.

Side Effects:
+ app.js modified to implement random image selection. Background used
only for larger display sizes.
+ Sticky footer added to site via changes to index.html and main.css.
Footer includes copyright info for background image.
+ Top navbar also made stick through changes to navbar.html.