Add interface for adding to camp and event list.
Adds a button to card directive which, when pressed, adds the entity of
the card to the appropriate list (camp or event). This was done in
service of the personalized list functionality.

Side Effects:
- Changes to card.directive.js to implement action.
- Changes to directive templates to provide button.
- Changes to Auth service to allow authentication (with login redirect
if necessary) on an action (like button press). Auth previously only
allowed this on state change.