Keep plan data in local storage.
Plan service now caches data in local storage. This was done so that
plan changes made in one window will be reflected in another window once
that window experiences a state change or reload. In the past this would
only happy on reload, which is probably not what the user expects (since
state change looks like reload to them).