ETL handles deletion without wiping index.
Changes the ETL service so that it does not delete the entire index
contents before performing the ETL operation. Instead, it indexes all
documents it retrieves from the burning man API, and deletes any
documents with IDs not loaded during the ETL process. This means the
user should never be able to tell when an ETL process is happening
(because the index won't magically disappear for a minute).