URL encode data in OAuth authentication request
Previously, AuthServerProvide would send a user's credentials as
application/x-www-form-urlencoded in a POST to the oauth token service.
The actual username and password, however, were not URL encoded before
transmission. Consequently, a bug was encountered where if users used
characters like '%' in their password, they could not login later.

This commit ensures the form data is URL encoded to avoid the problem.

The rememberMe field was also removed from the login controller, since
it is not used in the view and seems to serve no purpose in the
application. (maybe the rememberMe secret can be removed from yaml as
well?)