PlanService just makes ajax requests.
Nature of the plan service made it difficult to determine when a plan
was empty when dynamically updating the plan. Something funky was also
happening when logging in and out with multiple users on the same
machine. Now plan service just makes straight ajax requests and doesn't
try to cache anything. New endpoints to support its functionality.